skeleton-flask-api/requirements.txt
Flask
Flask-Cors
swagger-ui-py
python-dotenv
